refactor: extract helpers from large files
Extract utility functions and helper modules from oversized files:
- gameStore.ts (1608→1490 lines): Extract combatHelpers.ts and gameStateHelpers.ts
- RepoWatchTool.tsx (1077 lines): Extract repoWatchUtils.tsx
